+++ CHANGELOG.md
@@ -5,11 +5,20 @@ see the commit logs. The format is based on
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 0.5.3 - 2023-06-29
+
+### Added
+
+- Remote bootstrapping for Windows.
+- Temporary SSH connection argument.
+
+### Changed
+
+- GitHub role to install GitHub runner.
 
 ### Removed
 
-- Cuda, Obsidian, and Vagrant roles.
+- Cuda, DVC, Obsidian, and Vagrant roles.
 
 ## 0.5.2 - 2023-06-16